+++ libft/memmove.c
@@ -6,7 +6,7 @@
 /*   By: aperez-r <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/25 12:08:05 by aperez-r          #+#    #+#             */
-/*   Updated: 2024/09/25 16:50:17 by aperez-r         ###   ########.fr       */
+/*   Updated: 2024/09/30 12:12:40 by aperez-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 	unsigned char *d = (unsigned char*) dest;
 	unsigned const char *s = (unsigned const char*) src;
 	unsigned char temp[10];
-	int i;
+	size_t i;
 
 	i = 0;
 	while(i < n)
