+++ libft/strnstr.c
@@ -6,7 +6,7 @@
 /*   By: aperez-r <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/27 11:34:20 by aperez-r          #+#    #+#             */
-/*   Updated: 2024/09/27 12:30:14 by aperez-r         ###   ########.fr       */
+/*   Updated: 2024/09/30 12:14:20 by aperez-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if(needle[i] == '\0')
